Fixing Value Conversion Issues

+++ Passions/PASArtist.m
@@ -125,7 +125,7 @@
 		
 		// Setup a Timeout
 		 void (^timer)(void) = ^{
-			dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, kPASLastFmTimeoutInSec * NSEC_PER_SEC);
+			dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (NSInteger)(kPASLastFmTimeoutInSec * NSEC_PER_SEC));
 			dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
 				if ([task state] == NSURLSessionTaskStateRunning) {
 					// Task is running too long, cancel it
